requirements.txt
Flask==3.0.3
requests==2.31.0
python-dotenv==1.0.0
python-frontmatter==1.0.0
Markdown==3.6
